requirements.txt
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.92.0
google-cloud-storage 
python-docx
openai
Flask
twilio